# code/back/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Inventory Planning API
  description: |
    API dla systemu planowania zapasów kawy w biurowcach.
    
    System wspiera:
    - Optymalizację zamówień kawy z wykorzystaniem solvera CPLEX
    - Zarządzanie biurowcami i ich magazynami
    - Śledzenie zamówień i korekt
    - Autentykację użytkowników (JWT)
  version: 2.0.0
  contact:
    name: Zespół 2
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://api:8000
    description: Docker container
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token uzyskany z /auth/login

  # ==========================================
  # SCHEMAS
  # ==========================================
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Opis błędu
          example: "Resource not found"
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string

    # ---------- Auth ----------
    UserRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "securePassword123"
        first_name:
          type: string
          maxLength: 100
          example: "Jan"
        last_name:
          type: string
          maxLength: 100
          example: "Kowalski"
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Czas wygaśnięcia tokenu w sekundach
          example: 1800
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        first_name:
          type: string
          example: "Jan"
        last_name:
          type: string
          example: "Kowalski"
        role:
          type: string
          enum: [admin, manager, user]
          example: "user"
        office_id:
          type: integer
          nullable: true
          example: 1
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time

    # ---------- Office ----------
...
    OrderCreate:
      type: object
      required:
        - office_id
        - order_date
        - quantity_kg
        - unit_price
        - transport_cost
      properties:
        office_id:
          type: integer
          example: 1
        distributor_id:
          type: integer
          nullable: true
          example: 1
        order_date:
          type: string
          format: date
          example: "2026-01-15"
        delivery_date:
          type: string
          format: date
          example: "2026-01-16"
        quantity_kg:
          type: number
          format: float
          minimum: 0
          description: "x_{d,b,t} - zamówiona ilość [kg]"
          example: 25.0
        unit_price:
          type: number
          format: float
          minimum: 0
          description: "Cena jednostkowa [PLN/kg]"
          example: 45.0
        transport_cost:
          type: number
          format: float
          minimum: 0
          description: "Koszt transportu [PLN]"
          example: 50.0
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        optimization_request_id:
          type: integer
          nullable: true
          example: 1
        office_id:
          type: integer
          example: 1
        distributor_id:
          type: integer
          nullable: true
          example: 1
        order_date:
          type: string
          format: date
          example: "2026-01-15"
        delivery_date:
          type: string
          format: date
          example: "2026-01-16"
        quantity_kg:
          type: number
          format: float
          example: 25.0
        unit_price:
          type: number
          format: float
          example: 45.0
        transport_cost:
          type: number
          format: float
          example: 50.0
        total_cost:
          type: number
          format: float
          description: "Łączny koszt zamówienia [PLN]"
          example: 1175.0
        status:
          type: string
          enum: [planned, confirmed, delivered, cancelled]
          example: "planned"
        created_at:
          type: string
          format: date-time
        corrections:
          type: array
          items:
            $ref: '#/components/schemas/OrderCorrectionResponse'
...
    OrderCorrectionCreate:
      type: object
      required:
        - quantity_increase
        - quantity_decrease
      properties:
        quantity_increase:
          type: number
          format: float
          minimum: 0
          default: 0
          description: "r^+_{d,b,t} - zwiększenie zamówienia [kg]"
          example: 5.0
        quantity_decrease:
          type: number
          format: float
          minimum: 0
          default: 0
          description: "r^-_{d,b,t} - zmniejszenie zamówienia [kg]"
          example: 0.0
        reason:
          type: string
          example: "Zwiększone zapotrzebowanie z powodu konferencji"
    OrderCorrectionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        original_order_id:
          type: integer
          example: 1
        correction_date:
          type: string
          format: date-time
        quantity_increase:
          type: number
          format: float
          example: 5.0
        quantity_decrease:
          type: number
          format: float
          example: 0.0
        correction_cost:
          type: number
          format: float
          description: "Koszt korekty [PLN]"
          example: 25.0
        reason:
          type: string
          example: "Zwiększone zapotrzebowanie z powodu konferencji"

    # ---------- Optimization ----------
...
    SystemParameter:
      type: object
      properties:
        id:
          type: integer
          example: 1
        parameter_name:
          type: string
          example: "default_transport_cost"
        parameter_value:
          type: string
          example: "50.0"
        description:
          type: string
          example: "Domyślny koszt transportu [PLN]"
        updated_at:
          type: string
          format: date-time
    SystemParameterUpdate:
      type: object
      required:
        - parameter_value
      properties:
        parameter_value:
          type: string
          example: "75.0"

    # ---------- Legacy Predictions ----------
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Rejestracja nowego użytkownika
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Użytkownik utworzony
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email już istnieje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Błąd walidacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /auth/login:
    post:
      tags:
        - auth
      summary: Logowanie użytkownika
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Pomyślne logowanie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Nieprawidłowe dane logowania
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - auth
      summary: Pobranie danych zalogowanego użytkownika
      responses:
        '200':
          description: Dane użytkownika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------- Offices ----------
...
  /orders:
    get:
      tags:
        - orders
      summary: Lista zamówień z filtrowaniem
      parameters:
        - name: office_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [planned, confirmed, delivered, cancelled]
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        '200':
          description: Lista zamówień
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - orders
      summary: Ręczne utworzenie zamówienia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Zamówienie utworzone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Błąd walidacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /orders/{id}:
    get:
      tags:
        - orders
      summary: Szczegóły zamówienia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Szczegóły zamówienia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Zamówienie nie znalezione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/corrections:
    get:
      tags:
        - orders
      summary: Lista korekt dla zamówienia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista korekt
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderCorrectionResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Zamówienie nie znalezione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - orders
      summary: Utworzenie korekty zamówienia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCorrectionCreate'
      responses:
        '201':
          description: Korekta utworzona
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCorrectionResponse'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Zamówienie nie znalezione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Błąd walidacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  # ---------- Optimization ----------
...
  /settings:
    get:
      tags:
        - settings
      summary: Lista parametrów systemowych
      responses:
        '200':
          description: Lista parametrów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemParameter'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settings/{name}:
    put:
      tags:
        - settings
      summary: Aktualizacja parametru systemowego
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemParameterUpdate'
      responses:
        '200':
          description: Parametr zaktualizowany
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemParameter'
        '401':
          description: Brak autoryzacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Parametr nie znaleziony
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------- Legacy Predictions (kompatybilność wsteczna) ----------
...